refactor(framework): rename ArgumentParserNew to ArgumentParser
